requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
decorator==5.1.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
outcome==1.2.0
packaging==22.0
pluggy==1.0.0
PySocks==1.7.1
pytest==7.2.0
selenium==4.7.2
self==2020.12.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
